# docs/openapi.yaml
openapi: 3.0.3
info:
  title: travelu-api
  description: >
    API REST monolítica do travelu, organizada por domínio. Documenta exatamente
    o que `cmd/api/main.go` registra no mux: health, auth, categories, bookings,
    listings, reviews e trip-dates. Todas as rotas `/api/v1/**` exceto
    `/auth/*` exigem `Authorization: Bearer <access_token>`.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local (PORT default 8080)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/trip-dates:
    post:
      operationId: createTripDate
      tags: [trip-dates]
      summary: Cria uma janela de data reservável para um listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripDateRequest"
      responses:
        "201":
          description: Trip date criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      operationId: listTripDatesByListing
      tags: [trip-dates]
      summary: Lista as trip dates de um listing
      parameters:
        - name: listing_id
          in: query
          required: true
          description: Obrigatório — sem ele a API responde 400.
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Trip dates do listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripDate"
        "400":
          description: listing_id ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/trip-dates/{id}:
    get:
      operationId: getTripDateById
      tags: [trip-dates]
      summary: Busca uma trip date por ID
      parameters:
        - $ref: "#/components/parameters/TripDateId"
      responses:
        "200":
          description: Trip date encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDate"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT HS256 emitido por /api/v1/auth/login ou /refresh.
  parameters:
...
    TripDateId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: unauthorized, message: missing or invalid authorization token }
    ValidationError:
      description: Corpo inválido ou falhou uma regra de validação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: validation_error, message: "Key: 'CreateInput.Name' Error:Field validation for 'Name' failed on the 'required' tag" }
  schemas:
...
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
      required: [code, message]
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error"
      required: [error]

    # ---- auth / user ----
...
    CreateTripDateRequest:
      type: object
      description: end_date deve ser estritamente posterior a start_date (gtfield).
      properties:
        listing_id: { type: string, format: uuid }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        max_capacity:
          type: integer
          minimum: 1
          nullable: true
          description: Quando omitido, cai para a capacidade do próprio listing.
      required: [listing_id, start_date, end_date]
    TripDate:
      type: object
      properties:
        id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        max_capacity: { type: integer, nullable: true }
        current_bookings: { type: integer }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, listing_id, start_date, end_date, current_bookings, is_active, created_at, updated_at]